requirements.txt
anytree==2.8.0
git+https://github.com/cedadev/directory-tree#egg=directory_tree
python-dateutil==2.8.1
PyYAML==5.4.1
six==1.16.0
